Add royalty requirement to psy-kick's relevant stats
As far as I know, my mod only has two DLC-dependent features, the psyfocus boost on injecting psy-kick, and the psyfocus recovery stat on the psy-kick hediff. The former didn't actual cause errors if not locked, probably because RW is mono-library, but this feels like a 'better safe than sorry' scenario.