Rework How Buffs from Special Material Apparel Works
Adding an empty comp to every piece of apparel is horribly inefficient, so we really don't want to do that.

New functionality uses a new def which loads mappings of stuff and resuling hediffs into a static map which is leveraged in a harmony patch on the wear function to add the hediff when appropriate.

Thankfully this approach requires no new persistent data on the clothing, leather, pawn, or hediff, since the existing hediff comp to tie a hediff to apparel is sufficient.

Also makes this mod load before the meats optimization mod in preparation for a compatibility patch.